Passage Position is now a string; getStartingPassage() added
If a Twee3 file has position metadata, it will be read and transformed into a string value. This will also be written out when compiled into HTML.

The Story object now has a getStartingPassage() function that looks for the position of the Start passage, if it exists, or searches for the named passage from the Story metadata, if set.